@@ -4,17 +4,18 @@
   "version": "2.1.0",
   "type": "module",
   "scripts": {
-    "build": "ng build --configuration=production",
+    "build": "ng build",
     "ci": "yarn clean-dist && yarn clean-modules && yarn install --frozen-lockfile",
-    "clean": "yarn clean-dist && yarn clean-modules && rm yarn.lock",
+    "clean": "yarn clean-dist && yarn clean-modules && yarn clean-nx && rm yarn.lock",
     "clean-dist": "rm -rf dist",
     "clean-modules": "rm -rf node_modules",
+    "clean-nx": "rm -rf .nx/",
     "dev": "ng serve",
-    "pre-commit": "yarn lint && yarn prettier:check && yarn test:coverage",
+    "pre-commit": "yarn lint && yarn prettier && yarn test:coverage && yarn build",
     "prettier": "npx prettier --write .",
     "prettier:check": "npx prettier --check .",
     "test": "vitest run",
-    "test:dev": "vitest dev",
+    "test:dev": "vitest dev --reporter=basic --disable-console-intercept",
     "test:coverage": "vitest run --coverage",
     "lint": "ng lint",
     "lint:fix": "ng lint --fix",
@@ -30,25 +31,23 @@
   "dependencies": {
     "@analogjs/content": "^1.2.2",
     "@analogjs/router": "^1.2.2",
-    "@angular/animations": "^17.0.5",
-    "@angular/cdk": "^17.0.2",
-    "@angular/common": "^17.0.5",
-    "@angular/compiler": "^17.0.5",
-    "@angular/core": "^17.0.5",
-    "@angular/forms": "^17.0.5",
-    "@angular/platform-browser": "^17.0.5",
-    "@angular/platform-browser-dynamic": "^17.0.5",
+    "@angular/animations": "^17.2.0",
+    "@angular/cdk": "17.3.6",
+    "@angular/common": "^17.2.0",
+    "@angular/compiler": "^17.2.0",
+    "@angular/core": "^17.2.0",
+    "@angular/forms": "^17.2.0",
+    "@angular/platform-browser": "^17.2.0",
+    "@angular/platform-browser-dynamic": "^17.2.0",
     "@angular/platform-server": "^17.2.0",
-    "@angular/router": "^17.0.5",
+    "@angular/router": "^17.2.0",
     "@nx/angular": "~18.0.0",
-    "@vercel/node": "^3.0.27",
     "front-matter": "^4.0.2",
     "googleapis": "^135.0.0",
     "marked": "^5.0.2",
     "marked-gfm-heading-id": "^3.1.0",
     "marked-highlight": "^2.0.1",
     "mermaid": "^10.2.4",
-    "ngx-markdown": "^17.2.1",
     "nodemailer": "^6.9.13",
     "prismjs": "^1.29.0",
     "rxjs": "~7.5.6",
@@ -64,10 +63,11 @@
     "@angular-eslint/eslint-plugin-template": "17.3.0",
     "@angular-eslint/schematics": "17.3.0",
     "@angular-eslint/template-parser": "17.3.0",
-    "@angular/cli": "^17.0.5",
-    "@angular/compiler-cli": "^17.0.5",
+    "@angular/cli": "^17.2.0",
+    "@angular/compiler-cli": "^17.2.0",
     "@nx/vite": "~18.0.0",
     "@types/nodemailer": "^6.4.14",
+    "@types/prismjs": "^1.26.3",
     "@typescript-eslint/eslint-plugin": "7.2.0",
     "@typescript-eslint/parser": "7.2.0",
     "@vitest/coverage-v8": "^1.5.2",
