@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
 import { NotificationService } from '../../services/notification.service';
@@ -7,29 +6,34 @@ import { Notification, NotificationType } from '../../services/notification';
 @Component({
   selector: 'app-notification',
   standalone: true,
-  imports: [CommonModule],
   template: `
-    <div *ngIf="visibility" class="notification-container">
-      <div class="notification">
-        <div *ngIf="type === NotificationType.SUCCESS" class="icon icon-color-success">
-          <i class="ri-checkbox-circle-fill"></i>
-        </div>
-        <div *ngIf="type === NotificationType.FAILED" class="icon icon-color-failed">
-          <i class="ri-close-circle-fill"></i>
-        </div>
-        <div class="content">
-          <div class="title">{{ title }}</div>
-          <div class="message">{{ message }}</div>
-        </div>
-        <div class="close-button">
-          <i
-            (click)="hideNotification()"
-            (keyup)="hideNotification()"
-            class="ri-close-fill"
-            tabindex="0"></i>
+    @if (visibility) {
+      <div class="notification-container">
+        <div class="notification">
+          @if (type === NotificationType.SUCCESS) {
+            <div class="icon icon-color-success">
+              <i class="ri-checkbox-circle-fill"></i>
+            </div>
+          }
+          @if (type === NotificationType.FAILED) {
+            <div class="icon icon-color-failed">
+              <i class="ri-close-circle-fill"></i>
+            </div>
+          }
+          <div class="content">
+            <div class="title">{{ title }}</div>
+            <div class="message">{{ message }}</div>
+          </div>
+          <div class="close-button">
+            <i
+              (click)="hideNotification()"
+              (keyup)="hideNotification()"
+              class="ri-close-fill"
+              tabindex="0"></i>
+          </div>
         </div>
       </div>
-    </div>
+    }
   `,
   styleUrls: ['./notification.component.scss'],
 })
@@ -45,7 +49,6 @@ export class NotificationComponent implements OnInit, Notification {
   ngOnInit() {
     this.message$ = this.notificationService.getNotification();
     this.message$.subscribe((notificationReceived: Notification) => {
-      console.log({ notificationReceived });
       this.title = notificationReceived.title;
       this.type = notificationReceived.type;
       this.message = notificationReceived.message;
@@ -54,7 +57,6 @@ export class NotificationComponent implements OnInit, Notification {
       if (this.visibility) {
         setTimeout(() => {
           this.hideNotification();
-          console.log('visibility: ', this.visibility);
         }, 3000);
       }
     });
