@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import {
   AfterViewInit,
   ChangeDetectorRef,
@@ -8,102 +7,109 @@ import {
   OnInit,
   ViewChild,
 } from '@angular/core';
+import { Router } from '@angular/router';
 import { SearchService } from '../../services/search.service';
 import { SearchItem, SearchItemNames, SearchItemShortcuts } from './search-item';
-import { Router } from '@angular/router';
 import { ShortcutService } from '../../services/shortcut.service';
 
 @Component({
   selector: 'app-search',
   standalone: true,
-  imports: [CommonModule],
   template: `
-    <div id="{{ searchElementId }}" *ngIf="isSearchModuleVisible" class="c-hWUbGb">
-      <div class="c-iqTopT">
-        <div>
-          <input
-            #searchInput
-            type="text"
-            placeholder="Type a command or search…"
-            (input)="performSearch($event)"
-            class="c-ekIDDi"
-            autocomplete="off"
-            role="combobox"
-            spellcheck="false"
-            aria-expanded="true"
-            aria-controls="kbar-listbox"
-            aria-activedescendant="kbar-listbox-item-1" />
-          <div class="c-ertyu">
-            <div role="listbox" id="kbar-listbox" class="c-oeiotf">
-              <div class="c-title" *ngIf="generalCPItems.length > 0">
-                <div class="c-kcmNSe">General</div>
-              </div>
-              <div
-                *ngFor="let item of generalCPItems"
-                (click)="handleClick(item)"
-                (keyup)="handleClick(item)"
-                role="option"
-                aria-selected="true"
-                class="c-condee"
-                tabindex="0">
-                <div class="PJLV-iljSPlw-css">
-                  <div class="c-eSulSs">
-                    <div class="c-imgsize">
-                      <img
-                        alt="link"
-                        class="c-imgsize"
-                        aria-hidden="true"
-                        [src]="item.src"
-                        loading="lazy" />
-                    </div>
-                    <div class="c-fixGjY">
-                      <span>{{ item.label }}</span>
-                    </div>
+    @if (isSearchModuleVisible) {
+      <div id="{{ searchElementId }}" class="c-hWUbGb">
+        <div class="c-iqTopT">
+          <div>
+            <input
+              #searchInput
+              type="text"
+              placeholder="Type a command or search…"
+              (input)="performSearch($event)"
+              class="c-ekIDDi"
+              autocomplete="off"
+              role="combobox"
+              spellcheck="false"
+              aria-expanded="true"
+              aria-controls="kbar-listbox"
+              aria-activedescendant="kbar-listbox-item-1" />
+            <div class="c-ertyu">
+              <div role="listbox" id="kbar-listbox" class="c-oeiotf">
+                @if (generalCPItems.length > 0) {
+                  <div class="c-title">
+                    <div class="c-kcmNSe">General</div>
                   </div>
-                  <div aria-hidden="true" class="c-jpnQgQ">
-                    <kbd *ngFor="let shortcut of item.shortcut" class="c-ddlVgM">{{
-                      shortcut
-                    }}</kbd>
-                  </div>
-                </div>
-              </div>
-              <div *ngIf="pageCPItems.length > 0" class="c-title">
-                <div class="c-kcmNSe">Go to</div>
-              </div>
-              <div
-                *ngFor="let item of pageCPItems"
-                (click)="handleClick(item)"
-                (keyup)="handleClick(item)"
-                role="option"
-                aria-selected="true"
-                class="c-condee"
-                tabindex="0">
-                <div class="PJLV-iljSPlw-css">
-                  <div class="c-eSulSs">
-                    <div class="c-imgsize">
-                      <img
-                        alt="link"
-                        class="c-imgsize"
-                        aria-hidden="true"
-                        [src]="item.src"
-                        loading="lazy" />
-                    </div>
-                    <div class="c-fixGjY">
-                      <span>{{ item.label }}</span>
+                }
+                @for (item of generalCPItems; track item.name) {
+                  <div
+                    (click)="handleClick(item)"
+                    (keyup)="handleClick(item)"
+                    role="option"
+                    aria-selected="true"
+                    class="c-condee"
+                    tabindex="0">
+                    <div class="PJLV-iljSPlw-css">
+                      <div class="c-eSulSs">
+                        <div class="c-imgsize">
+                          <img
+                            alt="link"
+                            class="c-imgsize"
+                            aria-hidden="true"
+                            [src]="item.src"
+                            loading="lazy" />
+                        </div>
+                        <div class="c-fixGjY">
+                          <span>{{ item.label }}</span>
+                        </div>
+                      </div>
+                      <div aria-hidden="true" class="c-jpnQgQ">
+                        @for (shortcut of item.shortcut; track $index) {
+                          <kbd class="c-ddlVgM">{{ shortcut }}</kbd>
+                        }
+                      </div>
                     </div>
                   </div>
-                  <div aria-hidden="true" class="c-jpnQgQ">
-                    <kbd *ngFor="let shortcut of item.shortcut" class="c-ddlVgM">{{
-                      shortcut
-                    }}</kbd>
+                }
+                @if (pageCPItems.length > 0) {
+                  <div class="c-title">
+                    <div class="c-kcmNSe">Go to</div>
                   </div>
-                </div>
+                }
+                @for (item of pageCPItems; track item.name) {
+                  <div
+                    (click)="handleClick(item)"
+                    (keyup)="handleClick(item)"
+                    role="option"
+                    aria-selected="true"
+                    class="c-condee"
+                    tabindex="0">
+                    <div class="PJLV-iljSPlw-css">
+                      <div class="c-eSulSs">
+                        <div class="c-imgsize">
+                          <img
+                            alt="link"
+                            class="c-imgsize"
+                            aria-hidden="true"
+                            [src]="item.src"
+                            loading="lazy" />
+                        </div>
+                        <div class="c-fixGjY">
+                          <span>{{ item.label }}</span>
+                        </div>
+                      </div>
+                      <div aria-hidden="true" class="c-jpnQgQ">
+                        @for (shortcut of item.shortcut; track $index) {
+                          <kbd class="c-ddlVgM">{{ shortcut }}</kbd>
+                        }
+                      </div>
+                    </div>
+                  </div>
+                }
               </div>
             </div>
           </div>
         </div>
       </div>
-    </div>
+    }
   `,
   styleUrls: ['./search.component.scss'],
 })
@@ -170,8 +176,12 @@ export class SearchComponent implements OnInit, AfterViewInit {
         break;
 
       case SearchItemNames.HOME:
+        this.searchService.hideSearchModule();
+        this.router.navigate(['/']);
+        break;
+
       case SearchItemNames.ABOUT:
-      case SearchItemNames.ARTICLES:
+      case SearchItemNames.BLOG:
       case SearchItemNames.CONTACT:
       case SearchItemNames.PROJECTS:
         this.searchService.hideSearchModule();
@@ -213,10 +223,10 @@ export class SearchComponent implements OnInit, AfterViewInit {
         shortcut: [SearchItemShortcuts.ABOUT],
       },
       {
-        name: SearchItemNames.ARTICLES,
-        src: '/src/assets/svg/articles.svg',
-        label: `Articles`,
-        shortcut: [SearchItemShortcuts.ARTICLES],
+        name: SearchItemNames.BLOG,
+        src: '/src/assets/svg/blog.svg',
+        label: `Blog`,
+        shortcut: [SearchItemShortcuts.BLOG],
       },
       {
         name: SearchItemNames.PROJECTS,
