@@ -4,4 +4,4 @@ import { bootstrapApplication } from '@angular/platform-browser';
 import { AppComponent } from './app/app.component';
 import { appConfig } from './app/app.config';
 
-await bootstrapApplication(AppComponent, appConfig);
+bootstrapApplication(AppComponent, appConfig).catch((err) => console.error(err));
