@@ -14,11 +14,8 @@ export default defineConfig(({ mode }) => ({
   },
   plugins: [
     analog({
-      nitro: {
-        preset: 'vercel',
-      },
       prerender: {
-        routes: ['/home'],
+        routes: ['/', '/about', '/blog'],
       },
     }),
   ],
